Move writing functions to own script.
